requirements.txt
Flask>=3.0
opencv-python
python-dotenv
openai>=1.35
werkzeug>=3.0
